Organize all the files
