+++ src/main/java/Controller/UserSystemController.java
@@ -9,7 +9,7 @@ public class UserSystemController {
 
     private UserSystemController(UserManager um) { this.um = um; }
 
-    public void create(String phone_num, User user) { um.createUser(phone_num, user); }
+    public void create(String phone_num, User user) { um.addUser(phone_num, user); }
 
     public boolean verify(String phone_num, String enter_password) {
         return um.verifyUser(phone_num,  enter_password);
